+++ CHANGELOG.md
@@ -6,8 +6,9 @@ All notable changes will be documented in this file.
 
 ### Added
 
+* `NotSupported` exception (clean-up)
 * `CustomMetrics` (#8)
-* `NotImplemented` exception.
+* `NotImplemented` exception (clean-up)
 * `Error::CONDITIONAL_REQUEST_HAS_FAILED` and `Error::UNKNOWN_ORDER_ID` (#25)
 * `utils::was_order_received` and `utils::to_request_status` (#25)
 * `OrderAck::side` (#11)